Refactored code
1. Main functionality moved from main() so it can be called as
   a library.
2. Files templates modified to not use globals. This allows us to
   return memory as soon as we are done with the asset instead of
   leaking it.
3. Changed APIs to make it more convenient to list all files (instead
   of WalkDirs() )
4. ReadFile() does not need to decompress the entire filesystem to
   access one file.
