--- _example/gin/.gitignore
@@ -1,1 +1,0 @@
-myEmbeddedFiles
