--- _example/gin/b0x.yml
@@ -1,81 +1,0 @@
-# all folders and files are relative to the path 
-# where fileb0x was run at!
-
-# default: main
-pkg: myEmbeddedFiles
-
-# destination
-dest: "./myEmbeddedFiles/"
-
-# gofmt
-# type: bool
-# default: false
-fmt: true
-
-# compress files
-# at the moment, only supports gzip
-#
-# type: object
-compression:
-  # activates the compression
-  #
-  # type: bool
-  # default: false
-  compress: false
-
-  # valid values are:
-  # -> "NoCompression"
-  # -> "BestSpeed"
-  # -> "BestCompression"
-  # -> "DefaultCompression" or ""
-  #
-  # type: string
-  # default: "DefaultCompression" # when: Compress == true && Method == ""
-  method: ""
-
-  # true = do it yourself (the file is written as gzip compressed file into the memory file system)
-  # false = decompress files at run time (while writing file into memory file system)
-  #
-  # type: bool
-  # default: false
-  keep: false
-
-# ---------------
-# -- DANGEROUS --
-# ---------------
-# 
-# cleans the destination folder (only b0xfiles)
-# you should use this when using the spread function
-# type: bool
-# default: false
-clean: true
-
-# default: ab0x.go
-output: "ab0x.go"
-
-# [unexporTed] builds non-exporTed functions, variables and types...
-# type: bool
-# default: false
-unexporTed: false
-
-# [spread] means it will make a file to hold all fileb0x data
-# and each file into a separaTed .go file
-#
-# example:
-# theres 2 files in the folder assets, they're: hello.json and world.txt
-# when spread is activaTed, fileb0x will make a file: 
-# b0x.go or [output]'s data, assets_hello.json.go and assets_world.txt.go
-#
-#
-# type: bool
-# default: false
-spread: true
-
-# [lcf] log changed files when spread is active
-lcf: true
-
-# type: array of objects
-custom:
-  - files:
-    - "css/*"
-    - "templates/*"
