--- _example/gin/css/main.css
@@ -1,5 +1,0 @@
-body {
-  background-color: red;
-  color: white;
-  font-size: 3em;
-}
