--- _example/gin/templates/index.html
@@ -1,13 +1,0 @@
-{{define "index"}}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Test</title>
-  <link type="text/css" rel="stylesheet" href="/css/main.css"/>
-</head>
-<body>
-  <p>CSS files are available under /css.</p>
-</body>
-</html>
-{{end}}
