--- _example/simple/b0x.bat
@@ -1,1 +1,0 @@
-fileb0x b0x.json
