--- _example/simple/go.mod
@@ -1,22 +1,0 @@
-module example.com/foo/simple
-
-require (
-	github.com/BurntSushi/toml v0.3.1 // indirect
-	github.com/UnnoTed/fileb0x v0.0.0-20180803232343-279e1dc62738 // indirect
-	github.com/airking05/termui v2.2.0+incompatible // indirect
-	github.com/bmatcuk/doublestar v1.1.1 // indirect
-	github.com/karrick/godirwalk v1.7.3 // indirect
-	github.com/labstack/echo v3.2.1+incompatible // indirect
-	github.com/labstack/gommon v0.2.7 // indirect
-	github.com/maruel/panicparse v1.1.1 // indirect
-	github.com/mattn/go-colorable v0.0.9 // indirect
-	github.com/mattn/go-isatty v0.0.4 // indirect
-	github.com/mattn/go-runewidth v0.0.3 // indirect
-	github.com/mitchellh/go-wordwrap v1.0.0 // indirect
-	github.com/nsf/termbox-go v0.0.0-20180819125858-b66b20ab708e // indirect
-	github.com/valyala/bytebufferpool v1.0.0 // indirect
-	github.com/valyala/fasttemplate v0.0.0-20170224212429-dcecefd839c4 // indirect
-	golang.org/x/crypto v0.0.0-20180910181607-0e37d006457b // indirect
-	golang.org/x/net v0.0.0-20180921000356-2f5d2388922f
-	gopkg.in/yaml.v2 v2.2.1 // indirect
-)
