--- _example/simple/public/assets/data/exclude_me.txt
@@ -1,1 +1,0 @@
-i like turtles
