--- _example/simple/public/assets/data/test1.json
@@ -1,4 +1,0 @@
-{
-  "he": "llo",
-  "replace_test": "{world}"
-}
