--- _example/simple/public/assets/data/test2.json
@@ -1,3 +1,0 @@
-{
-  "email": "{EMAIL}"
-}
