--- _example/simple/public/secrets.txt
@@ -1,1 +1,0 @@
-jet fuel can't melt steel beams
