--- _example/simple/public/this is just some random stuff!.txt
@@ -1,1 +1,0 @@
-yeah
