--- _example/simple/reinstall.sh
@@ -1,4 +1,0 @@
-cd ../..
-go install
-cd _example/simple
-
