--- _example/simple/run
@@ -1,2 +1,0 @@
-go generate
-go run main.go
