--- _example/simple/run.bat
@@ -1,1 +1,0 @@
-go run main.go
