+++ go.mod
@@ -1,17 +1,22 @@
 module github.com/Velocidex/fileb0x
 
 require (
-	github.com/BurntSushi/toml v0.3.1
-	github.com/UnnoTed/fileb0x v1.1.4
-	github.com/bmatcuk/doublestar v1.1.1
+	github.com/BurntSushi/toml v1.4.0
+	github.com/bmatcuk/doublestar v1.3.4
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
 	github.com/gizak/termui/v3 v3.1.0
-	github.com/karrick/godirwalk v1.7.8
+	github.com/karrick/godirwalk v1.17.0
 	github.com/labstack/echo v3.2.1+incompatible
 	github.com/labstack/gommon v0.3.0 // indirect
+	github.com/mattn/go-runewidth v0.0.16 // indirect
+	github.com/mitchellh/go-wordwrap v1.0.1 // indirect
+	github.com/nsf/termbox-go v1.1.1 // indirect
+	github.com/rivo/uniseg v0.4.7 // indirect
 	github.com/stretchr/testify v1.4.0
 	golang.org/x/crypto v0.0.0-20200117160349-530e935923ad // indirect
 	golang.org/x/net v0.0.0-20190404232315-eb5bcb51f2a3
-	gopkg.in/yaml.v2 v2.2.2
+	gopkg.in/yaml.v2 v2.4.0
 )
 
 go 1.13
