+++ template/template_test.go
@@ -4,10 +4,10 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/UnnoTed/fileb0x/compression"
-	"github.com/UnnoTed/fileb0x/dir"
-	"github.com/UnnoTed/fileb0x/file"
-	"github.com/UnnoTed/fileb0x/updater"
+	"github.com/Velocidex/fileb0x/compression"
+	"github.com/Velocidex/fileb0x/dir"
+	"github.com/Velocidex/fileb0x/file"
+	"github.com/Velocidex/fileb0x/updater"
 	"github.com/stretchr/testify/assert"
 )
